トリミングとダウンロードをシーン単位のコンポーネントに分離
ImageSplitterScene をトリミング用の状態管理・イベント処理のみを持つ
オーケストレーターとし、UIは ImageTrimScene（画像選択〜切り抜き〜
「3分割する」まで）と ImageSplitResultScene（分割結果のダウンロード
一覧）に分離した。

- 「3分割する」完了後は columnImages の有無で表示を出し分け、
  トリミング側のUI（説明文・画像選択・Stage・縦横比ロック・ツール
  バー）は非表示になり、ダウンロード専用のシーンだけが表示される
- ダウンロードシーンの末尾に「もどる」を追加し、押すとトリミングの
  状態（画像・切り抜き範囲）を保持したままトリミングシーンに戻る
- 各シーンは props 経由でのみ状態・コールバックを受け取る見た目
  コンポーネントとして独立しており、あとから個別にデザインを当て
  やすい構成にした
